+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add support for building on Windows
+
+### Fixes
+
+- Also use `url` and `authToken` for upload command
+
 ## 0.0.2
 
 ### Features